requirements.txt
asgiref
Django
django-cors-header
djangorestframework
djangorestframwork-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
